Add Admin policies.

+++ app/controllers/api/v1/favorite_jobs_controller.rb
@@ -22,6 +22,7 @@ module Api
       end
 
       def destroy
+        authorize favorite
         if favorite.present?
           fav = favorite.destroy
           render status: 200, json: fav
